+++ CHANGELOG.md
@@ -46,6 +46,7 @@ lists they are listed as multiple entries.
 - Added initial partition support which allows you to connect to partition specified as a CLI argument
 - Added `listpartitions` CLI
 - Added Start and End Boundaries to ProgressBar increasing its Customizability
+- Added components to the theme. Components are user-defined reusable parts of TUI.
 
 ### Changed